Short to float conversions

+++ TR5Main/Objects/TR5/Entity/tr5_brownbeast.cpp
@@ -46,7 +46,7 @@ void ControlBrowsBeast(short itemNumber)
 	auto* item = &g_Level.Items[itemNumber];
 	auto* creature = GetCreatureInfo(item);
 
-	short angle  = 0;
+	float angle  = 0;
 
 	if (item->HitPoints <= 0)
 	{
