+++ TR5Main/Objects/TR5/Entity/tr5_chef.cpp
@@ -51,10 +51,10 @@ void ControlChef(short itemNumber)
 	if (!CreatureActive(itemNumber))
 		return;
 
-	short joint0 = 0;
-	short joint1 = 0;
-	short joint2 = 0;
-	short angle = 0;
+	float joint0 = 0;
+	float joint1 = 0;
+	float joint2 = 0;
+	float angle = 0;
 
 	auto* item = &g_Level.Items[itemNumber];
 	auto* creature = GetCreatureInfo(item);
