+++ TR5Main/Objects/TR5/Entity/tr5_cyborg.cpp
@@ -104,10 +104,10 @@ void CyborgControl(short itemNumber)
 		auto* creature = GetCreatureInfo(item);
 		auto* object = &Objects[item->ObjectNumber];
 
-		short angle = 0;
-		short joint2 = 0;
-		short joint1 = 0;
-		short joint0 = 0;
+		float angle = 0;
+		float joint2 = 0;
+		float joint1 = 0;
+		float joint0 = 0;
 
 		int x = item->Pose.Position.x;
 		int z = item->Pose.Position.z;
